% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brunel.R
\name{brunel}
\alias{brunel}
\title{Produce a Brunel visualization}
\usage{
brunel(brunel, data, width = 800, height = 600)
}
\arguments{
\item{brunel}{the brunel script}

\item{data}{the data frame containing the data
width the width in pixels for the resultingvisualization
height the heigh in pixels for the resulting visualization}
}
\description{
Given a brunel script and a supplied data frame, renders the visualization.
The BRUNEL_SERVER environment variable must be set to the location of the
Brunel web service.
}

